fix command bug
