+++ app/bootstrap.php
@@ -6,6 +6,16 @@
  * @author    Ivanis Kouamé <ivanis.kouame@smile.fr>
  * @copyright 2015 Smile
  */
+
+$loader = require_once __DIR__.'/../vendor/autoload.php';
+
+use Silex\Provider\FormServiceProvider;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Silex\Provider\SerializerServiceProvider;
+use Silex\Provider\TranslationServiceProvider;
+use Symfony\Component\Translation\Loader\YamlFileLoader;
+use Doctrine\Common\Annotations\AnnotationRegistry;
 use Doctrine\Common\Cache\ArrayCache;
 use Doctrine\ORM\Tools\Console\ConsoleRunner;
 use Doctrine\ORM\Tools\Setup;
@@ -43,6 +53,12 @@ use Dflydev\Silex\Provider\DoctrineOrm\DoctrineOrmServiceProvider;
 use HiPay\Wallet\Mirakl\Integration\Handler\MonologDBHandler;
 
 include dirname(__FILE__).'/../vendor/erusev/parsedown/Parsedown.php';
+
+AnnotationRegistry::registerLoader(array($loader, 'loadClass'));
+
+$app          = new Silex\Application();
+$app['debug'] = true;
+
 
 $paths = array(
     join(DIRECTORY_SEPARATOR, array(__DIR__, "..", "src", "Entity"))
@@ -198,3 +214,63 @@ $cashoutProcessor = new CashoutProcessor(
 
 $notificationHandler = new NotificationHandler($eventDispatcher, $logger, $operationRepository, $vendorRepository,
                                                $logVendorRepository, $apiFactory, $logOperationsRepository);
+
+$app->register(new Silex\Provider\SessionServiceProvider());
+
+$app->register(new Silex\Provider\TwigServiceProvider(),
+               array(
+    'twig.path' => __DIR__.'/../views',
+    'twig.options' => array(
+        'cache' => __DIR__.'/../var/cache',
+    ),
+));
+
+$app->register(new Silex\Provider\UrlGeneratorServiceProvider());
+
+$app->register(new Silex\Provider\ServiceControllerServiceProvider());
+
+$app['twig'] = $app->share($app->extend('twig',
+                                        function($twig, $app) {
+        $twig->addFunction(new \Twig_SimpleFunction('asset',
+                                                    function ($asset) use ($app) {
+            return sprintf('%s/%s', trim($app['request']->getBasePath()), ltrim($asset, '/'));
+        }));
+        return $twig;
+    }));
+
+$app->register(new Silex\Provider\ValidatorServiceProvider());
+
+$app->register(new Silex\Provider\TranslationServiceProvider(), array(
+    'translator.domains' => array(),
+));
+
+$app->register(new Silex\Provider\HttpCacheServiceProvider(),
+               array(
+    'http_cache.cache_dir' => __DIR__.'/../var/cache/',
+    'http_cache.esi' => null,
+));
+
+$app->register(new FormServiceProvider());
+
+$app->register(new SerializerServiceProvider());
+
+// set multilanguage support
+$app->register(new TranslationServiceProvider(), array(
+    'locale_fallbacks' => array('en'),
+));
+
+$app['translator'] = $app->share($app->extend('translator',
+                                              function($translator, $app) {
+        $translator->addLoader('yaml', new YamlFileLoader());
+        $translator->addResource('yaml', __DIR__.'/../app/locales/en.yml', 'en');
+        $translator->addResource('yaml', __DIR__.'/../app/locales/fr.yml', 'fr');
+
+        return $translator;
+    }));
+
+$app['translator']->setLocale($parameters['dashboard.locale']);
+$app->before(function (Request $request) use ($app) {
+    $app['twig']->addGlobal('active', $request->get("_route"));
+});
+
+return $app;
