+++ bin/console
@@ -6,7 +6,7 @@
  * @author    Ivanis Kouamé <ivanis.kouame@smile.fr>
  * @copyright 2015 Smile
  */
-require_once __DIR__ . '/../web/index.php';
+
 
 use Doctrine\ORM\Tools\Console\ConsoleRunner as ORMConsoleRunner;
 use HiPay\Wallet\Mirakl\Integration\Command\Vendor\ProcessCommand as VendorProcessCommand;
@@ -17,6 +17,8 @@ use HiPay\Wallet\Mirakl\Integration\Command\Wallet\BankInfosCommand;
 use HiPay\Wallet\Mirakl\Integration\Command\Wallet\ListCommand;
 use HiPay\Wallet\Mirakl\Integration\Command\Log\LogVendorsRecoverCommand;
 use Symfony\Component\Console\Application;
+
+$app = require_once __DIR__.'/../app/bootstrap.php';
 
 $console = new Application("Hipay Mirakl Connector Console");
